+++ src/types.ts
@@ -1,0 +1,7 @@
+export interface MovieData {
+  id: number
+  poster_path: string
+  title: string
+  release_date: string
+  vote_average: number
+}
